moved login and register to auth folder
