+++ src/components/auth/Register.js
